+++ src/openms/source/ANALYSIS/MAPMATCHING/ConsensusMapNormalizerAlgorithmQuantile.cpp
@@ -74,7 +74,7 @@ namespace OpenMS
       vector<double> sorted = feature_ints[i];
       std::sort(sorted.begin(), sorted.end());
       vector<double> resampled(largest_number_of_features);
-      resample(sorted, resampled, largest_number_of_features);
+      resample(sorted, resampled, static_cast<UInt>(largest_number_of_features));
       resampled_sorted_data.push_back(resampled);
     }
 
@@ -93,7 +93,7 @@ namespace OpenMS
     for (Size i = 0; i < number_of_maps; ++i)
     {
       vector<double> ints;
-      resample(reference_distribution, ints, feature_ints[i].size());
+      resample(reference_distribution, ints, static_cast<UInt>(feature_ints[i].size()));
       normalized_sorted_ints[i] = ints;
     }
